refactor(T9298): migrate history-adapters + structured-output from CompletionResult to backend.ts SSoT + delete legacy-types.ts
- history-adapters.ts: import CompletionResult from ./backend.js (canonical source)
- structured-output.ts: import CompletionResult from ./backend.js
- index.ts: replace legacy-types re-exports with backend.ts re-exports; drop completionResultToResponse (no external consumers); add BackendCallParams to public surface
- llm-layer.test.ts: update imports to ../backend.js
- deleted packages/core/src/llm/legacy-types.ts (was a shim with zero justified consumers)

Sub-task B (planAttempt): SKIP — active consumer in api.ts:321
Sub-task C (debt sweep): legacy-types was the only LLM-scope shim file; all other @deprecated hits in repo are in unrelated subsystems (hooks, nexus, caamp) with active consumers

Gate: grep legacy-types|CompletionResult outside backend/index/consumers → 0 hits
Gate: 60/60 llm-layer tests pass
