+++ packages/core/src/llm/history-adapters.ts
@@ -9,7 +9,7 @@
  * @epic T1386
  */
 
-import type { CompletionResult } from './legacy-types.js';
+import type { CompletionResult } from './backend.js';
 
 /** Protocol for history adapters. */
 export interface HistoryAdapter {
