+++ packages/core/src/llm/index.ts
@@ -12,6 +12,15 @@
 export type { CleoLlmCallParams } from './api.js';
 // Public entrypoint
 export { cleoLlmCall } from './api.js';
+// Backend types — canonical CompletionResult / ProviderBackend shapes used by request-builder, history-adapters, structured-output
+export type {
+  BackendCallParams,
+  CompletionResult,
+  ProviderBackend,
+  StreamChunk,
+  ToolCallResult,
+} from './backend.js';
+export { makeCompletionResult } from './backend.js';
 // Backends (for direct use / custom wiring — AnthropicBackend migrated to transports/anthropic.ts)
 export type { GeminiCacheHandle } from './caching.js';
 // Caching
@@ -78,14 +87,6 @@ export {
   GeminiHistoryAdapter,
   OpenAIHistoryAdapter,
 } from './history-adapters.js';
-// Legacy backend type shims — still used by history-adapters.ts (HistoryAdapter.formatAssistantToolMessage)
-export type {
-  CompletionResult,
-  ProviderBackend,
-  StreamChunk,
-  ToolCallResult,
-} from './legacy-types.js';
-export { completionResultToResponse, makeCompletionResult } from './legacy-types.js';
 export type { ModelMetadata } from './model-metadata.js';
 // Model metadata — context window resolution (T9264 / T-LLM-CRED Phase 3)
 export {
